print error message and exit when gerbv fails
As all output of the command was silenced and further there was no check
of the exit code, in the case of gerbv failing there would simply be no
output file which results in a hard to understand error of PIL
complaining that no png file exists.
Switched from subprocess.call() to subprocess.run() with the check
option set to True, which will result in an exception in case the
command fails. In case this exception is raised an error message will be
printed, also indicating that perhaps gerbv is simply not installed
(probably the most common mode of failure). As the png files are
required for the further execution of the program, in case of the
exception the program will exit immediately.